+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the "shell-preview" extension will be documented in this 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+- Improve output with exit code and stderr
+- Terminate commands if not finished after a timeout (default 20s)
 
 ## [0.0.3] - 2026-05-15